refactory: remove libgui & add simulate.c
