--- lib/libgui/gui_open_window.c
@@ -1,19 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   gui_open_window.c                                  :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: yonshin <yonshin@student.42.fr>            +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/06/05 06:16:35 by yonshin           #+#    #+#             */
-/*   Updated: 2023/06/05 06:16:52 by yonshin          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libgui.h"
-
-void	gui_open_window(t_gui_setting *g)
-{
-	mlx_loop(g->mlx);
-	mlx_terminate(g->mlx);
-}
