--- lib/libgui/gui_setting.c
@@ -1,53 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   gui_setting.c                                      :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: yonshin <yonshin@student.42.fr>            +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/06/05 05:22:39 by yonshin           #+#    #+#             */
-/*   Updated: 2023/06/07 06:28:46 by yonshin          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libgui.h"
-
-static void	_init_gui_setting(t_gui_setting *g, char *title)
-{
-	g->mlx = NULL;
-	g->image = NULL;
-	g->title = title;
-	g->focal_length = DEFAULT_FOCAL_WIDTH;
-	g->max_depth = DEFAULT_MAX_DEPTH;
-	g->ignore_delta = DEFAULT_IGNORE_DELTA;
-	g->sample_per_pixel = DEFAULT_SAMPLE_PER_PIXEL;
-}
-
-static int	_replace_image(t_gui_setting *g, int width, int height)
-{
-	if (g->image != NULL)
-		mlx_delete_image(g->mlx, g->image);
-	g->image = mlx_new_image(g->mlx, width, height);
-	if (g->image == NULL)
-	{
-		mlx_close_window(g->mlx);
-		return (ft_error(__func__, __FILE__, __LINE__, 0));
-	}
-	if (mlx_image_to_window(g->mlx, g->image, 0, 0) == -1)
-	{
-		mlx_close_window(g->mlx);
-		return (ft_error(__func__, __FILE__, __LINE__, 0));
-	}
-	return (EXIT_SUCCESS);
-}
-
-int	gui_setting(t_gui_setting *g, int width, int height, char *title)
-{
-	_init_gui_setting(g, title);
-	g->mlx = mlx_init(width, height, g->title, true);
-	if (g->mlx == NULL)
-		return (ft_error(__func__, __FILE__, __LINE__, 0));
-	if (_replace_image(g, width, height))
-		return (ft_error(__func__, __FILE__, __LINE__, 0));
-	return (EXIT_SUCCESS);
-}
